mysql folder refactor
