+++ mysql/src/models/user.model.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-console */
 /* eslint-disable func-names */
 /* eslint indent: 1 */
-const dbConn = require('../../mysql_config/db.config');
+const dbConn = require('../config/db.config');
 
 // eslint-disable-next-line
 //User Object create
